refactor: move to cli.py
